+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix bug in Participatory Budgeting where removing ideas in the selection was stored wrongly.
 * Fix incorrect closing tag in slider widget
 * ideas-on-map fixes and extensions
+* Fixes on choices-guide result forms
 
 ## v0.12.0 (2020-01-27)
 * Allow for running multiple sites on subdirectories